backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==4.2.17
django-allauth==65.3.0
django-cors-headers==4.6.0
django-filter==24.3
django-resized==1.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
